requirements.txt
gradio
python-dotenv
langchain-google-genai
langchain-core
langgraph
pandas